requirements.txt
azure-functions
azure-storage-blob>=12.20.0
Jinja2>=3.1.4
playwright>=1.47.0
